pyproject.toml
[project]
name = "ragent"
version = "0.1.0"
description = "Distributed RAG Agent System (Phase 1 skeleton)"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "elasticsearch-haystack>=4.0",
    "fastapi>=0.115.0",
    "haystack-ai>=2.0",
    "httpx>=0.27.0",
    "langdetect>=1.0.9",
    "minio>=7.2.0",
    "nltk>=3.9.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-sdk>=1.25.0",
    "prometheus-client>=0.21.0",
    "pymysql>=1.1.3",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0",
    "taskiq>=0.12.2",
    "taskiq-redis>=1.2.2",
    "uuid-utils>=0.9",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "bandit>=1.9.4",
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.5",
    "safety>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ragent"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q --tb=short"

[dependency-groups]
dev = [
    "fakeredis>=2.35.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
    "testcontainers[elasticsearch,mariadb,minio,redis]>=4.14.2",
]
